KPluginFactory: remove argumented contructor and componentName()
The Qt plugin system does not pass any arguments to its root object, so
the component name will never get set in normal use.  So there's no
point having a method to return it, or in accepting it in the
constructor.

This was the oddly-ported version of a method that returned a
KComponentData, and did not exist in kdelibs 4, hence it is being
removed rather than deprecated.

REVIEW: 117074