kmimeassociationstest: do not use KDE apps in test
Looking for applications with a kde4- prefix will not be useful in the
long term, as those applications will disappear as people move away from
kdelibs4. However, simply stripping the kde4- prefix also does not work
because you can set a mime type assocation for "kwrite.desktop", and if
the KF5-based KWrite (which has the desktop ID "kwrite.desktop") is not
installed but the kdelibs4-based one (which has the desktop ID
"kde4-kwrite.desktop") is installed, the mime type will be associated
with kde4-kwrite.desktop, and the test will fail.

To avoid this issue, we avoid using KDE applications in the tests.

The prefixing stuff is explicitly tested with some fake application
desktop files. We still try adding and removing associations for real
applications (that may or may not be installed) to test cascading.

REVIEW: 117475