Fix linker errors with MSVC.
KSycocaEntry has its copy ctor disabled with Q_DISABLE_COPY. For some
reason MSVC emits a copy ctor for KService (a subclass), which calls the
copy ctor of KSycocaEntry. This compiles, because KService is a friend of
KSycocaEntry, so the copy ctor is accessible despite being private.
However, it fails to link because the private copy ctor is not defined
anywhere. The same happens with the assignment operator.

Fixed by disabling copy on KService too, so that the compiler doesn't try
to emit a copy constructor or assignment operator for it.

REVIEW:117680