Fix KPluginInfo::dependencies() when constructing from a .desktop file
Previously it would treat X-KDE-PluginInfo-Depends= (with no value)
as a string list with one element (the empty string). When constructing
from a KService::Ptr using the same file we get an empty string list as
one would expect.

This makes the behaviour of the KPluginInfo(QString) constructor match
the KPluginInfo(KService::Ptr) version.

REVIEW: 119936