Improve StringList property handling
Makes it possible to interpret properties from plugins that expose
properties correctly in the json.
Currently it checks whether it's a stringlist value according to the
servicetype (good) and splits the ','.

What this patch does is to allow plugins to use json lists to specify
string list properties. For example:
    "X-KDevelop-Interfaces": [ "IBuildSystemManager" ]

REVIEW: 122514