Make ksycocathreadtest more reliable: lastModified() doesn't have milliseconds
so we need to wait until the beginning of the next second before
modifying ksycoca again. Let's hope nobody modifies it twice in the
same second in real life. I think the bug is QFileInfo::lastModified(),
the information about milliseconds does exist in the filesystem, I'll
look into fixing that.