KSycoca: check timestamps and run kbuildsycoca if needed. No kded needed anymore.
If 1.5s have passed since the last time we checked, we compare the mtime of
the directories (12 dirs on my system) with the timestamp stored in the
ksycoca database (which indicates that all changes prior to that time
are in the DB).

Note that we only check the mtime of dirs, not files. Therefore manually
editing an installed .desktop file will require touching the dir, or
running kbuildsycoca5.

REVIEW: 124876