Remove code for --checkstamps and --nocheckfiles.
(the options are kept to avoid breaking hypothetic callers, but do nothing anymore)

--checkstamps was a timestamp comparison to find out if kbuildsycoca should
do nothing. Let's not just call it in that case ;) A similar timestamp check
is in ksycoca.cpp since recently.

--nocheckfiles was an experimental and marked dangerous option,
to speed up KDE startup by only checking the global header (lang and dirs),
but not checking for modified desktop files. I guess the idea was for
binary packages to always ensure that ksycoca is updated. That's a very
long term plan with a fdo desktop-file-index, but not really applicable until then
(what if you install something from sources, on a distro set up that way...).

REVIEW: 125108