KBuildSycoca: replace dbus-based locking with a lock file.
This is a more traditional way to protect concurrent access to a file,
and it will (more easily) allow to have one lock file per sycoca file
when we have LANG and DIRS in the filename.

It doesn't exactly remove the dependency on DBus though, since we still
use it for the notification after the rebuild.

As an additional improvement, if a process has to wait for another
to rebuild sycoca, when it can finally get the lock file, it does a quick
mtime check and if there were no additional changes compared to the
time of the last check (as stored in ksycoca's global header) then there
is nothing to do, we can do an early return.

REVIEW: 125152