KSycoca: replace DBus notification with file watching using KDirWatch.
The list of changed resources is not longer available, so we fake it
by passing all 4 resources. No big deal, it just removes a small optimization
in some apps, and anyway the "apps" resource was emitted in all cases for the
last 8 years anyway (vfolder bug that was never fixed).
-> un-deprecate the old databaseChanged signal without the resource list ;)

KDirWatch beats QFileSystemWatcher once again, after much fighting with QFSW
with one failing unittest (KServiceTest::testThreads), I switched to KDirWatch
and it worked immediately...

REVIEW: 125185