KSycoca: rebuild sycoca in-process in most unittests as well.
This tests what happens now in applications, in order to catch any possible
issues with the in-process rebuild.
On the other hand, I left the QProcess usage in ksycocatest, so that this
one ensures the kbuildsycoca5 binary still works.

This change caught a nasty code path where the sycoca-header-check
during rebuild would actually trigger a rebuild, which would deadlock
on the lockfile. This bug was already in 5.14.0 but there the symptom
was harder to spot due to the process separation (constant rebuild, bug 352787).

REVIEW: 125272