KBuildSycoca: remove writing of the ksycoca5stamp file.
The code for reading it was already removed as part of b0c8fd8e64f,
the removal of the --checkstamps option.

The feature still exists anyway, a timestamp of "at least everything until
that time is in the DB" is in the sycoca header, and it used by the timestamp
check in ksycoca.cpp. This was an unnecessary separate file.

REVIEW: 125274