KSycoca: change DB filename to include language and sha1 of the dirs it's built from.
This will prevent sycoca-rebuild ping-pong if two apps with different settings
would share the same file (and keep finding that it's wrong for them),
and it fixes Albert's bug that "LANG=de kcmshell5 --list" doesn't show German
translations for the strings coming from desktop files.

To avoid migration issues, the old filename ksycoca5 is still provided, as a hardlink
so that mmap works.

REVIEW: 125279
BUGS: 340731, 335780
FIXED-IN: 5.15