[autotest] make sure kbuildsycoca has something to do.
If it says "Database is up to date" then the db won't be modified
and the filewatcher won't emit the signal.

This worked locally because app desktop files lead the vfolder code into
the old bug that the dict is never empty at the end so it thinks "apps" were modified.
But the CI has a cleaner env so this doesn't happen there.