KSycoca: store mtime for every source dir, to detect changes.
The old solution was to compare the mtime of the source dirs with
a stored timestamp of "just before listing them". However this assumes
that time always moves forward, and some people have a phone that also acts
as a time machine (a limited one, with a single possible "destination", 01/01/1970).

Well, this new solution should also avoid problems with DST making time
go back 1h once a year; or problems when changing timezones (although we
could also just use UTC everywhere for these issues).

BUG: 353036
REVIEW: 125497