Fix associated applications for mimetypes with uppercase characters in the name
As the mimetype name is stored in lowercase in the dictionary, we need
to convert it to lowercase in the dictionary check in
KMimeTypeFactory::serviceOffersOffset() as well.

Otherwise this check always fails if the name contains uppercase
characters, and the list of associated applications will always be
empty.

Problem reported in openSUSE's bugzilla:
http://bugzilla.opensuse.org/show_bug.cgi?id=980046

REVIEW: 128082