autotests: check that mimetypes defined by kcoreaddons are available
Clearly they are not, on Windows, in CI (this commit just makes the error
clearer).

Was a solution found for the problem of finding mimetypes (using
GenericDataLocation) in custom prefixes? I wasn't able to get QT_DATA_DIRS into
Qt, but this is another case where it would have helped...

CCMAIL: kde-windows@kde.org