[KService] add workingDirectory(), deprecate path().
Summary:
Reading code like service->path() is confusing, because it looks like
the path to the .desktop file (which is in fact entryPath(), most
likely because path() was already taken).

The key in the desktop file is called Path (and that won't change),
but in the C++ API, workingDirectory() is much clearer.

Test Plan: Builds

Reviewers: vkrause

Reviewed By: vkrause

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23939