[KService] Add copy constructor.
Summary:
I need this in KIO, where a (soon deprecated) method takes a KService by const
ref, but I want to implement a KJob behind the scenes, which means starting
delayed, so I need to clone that KService object. The job will take a
KService::Ptr to avoid needing this in future code.

Some context of reviewers: this is all based on QExplicitlySharedDataPointer.

Test Plan: Builds, new tests pass

Reviewers: vkrause, sitter, davidedmundson, apol

Reviewed By: vkrause

Differential Revision: https://phabricator.kde.org/D23940