KSycoca: avoid database rebuild if XDG_CONFIG_DIRS contains duplicates
Similar to ffbfbba051b29be371f7382536359107cf551d8b.

We compare  extraFiles.keys() to KBuildSycoca::factoryExtraFiles() to
see if there are any files added.

As extraFiles.keys is from a map, duplicates are implicitly removed.
This is not true for the stringlist.

Duplicates can arise if a user has the same directory twice in
XDG_CONFIG_DIRS or in XDG_CURRENT_DESKTOP.