Allow NotShowIn=KDE apps, listed in mimeapps.list, to be used
It was unexpected for users that selecting e.g. Gnome FileRoller in
mimeapps.list for a given mimetype would still not make it associated
to that mimetype when in Plasma. NotShowIn is documented to be about
"displaying" applications (in the K menu and context menus), not about
completely forbidding users from associating such apps with a given
mimetype.

Note that if one wants different file associations per desktop, that's
supported with gnome-mimeapps.list and kde-mimeapps.list.

The fix is to check showInCurrentDesktop() at cache creation time
when collecting .desktop files, but to not check it for mimeapps.list entries.

The trick to make this work was that we still need to process ONE
servicetype for such "hidden" desktop files: the "Application"
servicetype (internal detail which can hopefully go away in KF6).

BUG: 427469
FIXED-IN: 5.76