KPluginInfo: Fix mimetypes that are wildcards not appearing in MimeTypes json value
The kcoreaddons conversion code does not check the validity of mime types at all.
For example in the thumbnailers this causes mime types from the KServiceTypeTrader code
path to only appear in the ServiceTypes.

In the long run we are getting rid of KPluginInfo anyways, see https://phabricator.kde.org/T12178.
However we need to keep it around during KF5 for all the compatibility code paths.