Emit deprecation warning for KServiceType class
Deprecating this with the visibility guards will be a mess, because it is used quite often internally.

Instead we emit a deprecation warning for the external users and make it a implementation Detail in KF6.
This way we can clean it up at any time during the KF6 lifetime (if we don't need it anymore).

Looking at the usages I only see one in a KIO test, though it seems not that useful IMHO.

Task: https://phabricator.kde.org/T12181