QVariant(QVariant::Type) constructor is explicit in Qt6
This fixes KApplicationTraderTest::testTraderConstraints():
 - X-KDE-Version has the value "5.56" (string) that is supposed to be converted
   to double
 - when calling KService::property() the code ends up calling
   KConfigGroup::convertToQVariant(), when passing the arg "t" (QVariant::Type)
   no implicit conversion happens, and the QVariant that Kcfg::convertToQVariant()
   gets is of type QVariant::UInt (QVariant::Double has value 6, which is
   detected as an uint in convertToQVariant())

Explicitly constructing a QVariant works in Qt5 and 6, so use that.

Now all unit test pass with Qt6.

This code will need to be converted to use QMetaType::Type instead of the
deprecated QVariant::Type, but that's a bigger change, one that we'll have
to do in the whole KF stack starting from tier 1 all the way up. Should be
OK as the QVariant::Type enum values are almost 1:1 with QMetaType::Type. So
it should work in Qt5 too hopefully.