kbuildsycoca: Ignore last modified time when set to UNIX Epoch
On some systems (i.e. Fedora Kinoite), all files in /usr have a last
modified timestamp of 0 (UNIX Epoch). In this case, always assume the
file as been changed.

We only get the current time once to speed up execution as we don't need
to be accurate in this case.

BUG: 442011

See:
- https://bugzilla.redhat.com/show_bug.cgi?id=2021087
- https://discussion.fedoraproject.org/t/kinoite-installation-with-utc-timezone-breaks-kde/34293/8