Allow accessing custom properties of KServiceActions
It is possible to define custom properties for actions in .desktop files.

We use this for example in kglobalaccel for specifying per-action shortcuts.

Currently there is no way to read these properties from a KServiceAction though.

Add a property() method modeled after KService::property() to allow this.

The properties need to be saved in the syscoca database.
We can't really add any fields to the KServiceAction serialization without breaking the binary format though.
Fortunately there's the currently unused m_data field which we can use for this purpose