Parse actions last when creating a service
parseActions creates the KServiceAction objects, which get passed a clone
of the service.

By the time the clone happens the service should be fully parsed, otherwise
the clone will be incomplete.

Delay the parsing of the actions to the end to make sure everything else is
parsed

(cherry picked from commit 3b135fcc20e3785ccbbd1bb90725beb30f5e6510)