Fix preferred apps order for multiple level mime inheritance
We should prefer more specific mime types to less specific for that
reason offers are sorted by inheritance level first.
When constructing the offers for inherited types, the associated
services for parent types are queried. However these can already
contain services that are associated with a parent type of the
parent type as the offers are build from the bottom-up starting
with the least specific mime type.
For example when constructing offers for a mime-type that inherits
text/plain, the returned offers for text/plain will already contain
services associated with application/octet-stream.
By incrementing the inheritance levels instead of overwriting the
relative ordering of inherited mime types is preserved.
BUG:475200