kservice: remove InitialPreference
Since this was a KDE specific feature it caused problems with
interoperability. Applications based on other XDG implementations would
entirely ignore it and by extension have a different idea of what
application should be used to open a given mimetype. Instead we now rely
only on the XDG provided facilities to manage preference. Specifically
plasma-desktop is meant to ship a kde-mimeapps.list that encodes the
Plasma default preferences and distributions are meant to ship
*mimeapps.list to encode sensible preferences beyond that.

BUG: 305136