Remove disallowed services when queryByMimeType is called
Disallow services such as xdg-open and kde-open for opening
a specific mimetype.

Some applications set their file mimetypes to open with "xdg-open" which
then
causes it open in "kde-open", which then ends up calling itself
constantly and endlessly, spiking CPU.

Instead, this helps to catch that early on and return a null service, so
other apps can handle it.