Don't warn for empty layout without merge tag
The menu spec says

Each <Layout> or <DefaultLayout> element shall have exactly one <Merge type="all">
element or it shall have exactly one <Merge type="files">
and exactly one <Merge type="menus"> element.

but it also says

An exception is made for a completely empty <Layout> element which may
be used to indicate that the default-layout should be used instead.

So don't show the warning when the layout is empty.

This can happen when removing all entries of a category in kmenuedit